VM parsing is its own file
